requirements.txt
luma.oled
psutil
pillow